Fix Db to support tracking column overrides, defaults for SQLite

+++ adapters/sqlite.rb
@@ -30,15 +30,17 @@ module Rake
         table_exists?(TRACKING_TABLE_NAME)
       end
 
+      def self.tracking_table_columns
+        super.merge({
+          :time => {:data_type => :timestamp}
+        })
+      end
+
       def self.set_up_tracking
-        create_table TRACKING_TABLE_NAME, nil, "
-          (
-            relation_name text,
-            relation_type text,
-            operation text,
-            time timestamp
-          )
-        ", false
+        column_definitions = tracking_table_columns.map do |col,col_defn|
+          col.to_s + ' ' + col_defn[:data_type].to_s
+        end.join(', ')
+        create_table TRACKING_TABLE_NAME, nil, " (#{column_definitions})", false
       end
 
       def self.tear_down_tracking
@@ -58,11 +60,36 @@ module Rake
       end
 
       def self.truncate_table table_name
+        return if table_name.casecmp(TRACKING_TABLE_NAME) == 0
         Db.execute "delete from #{table_name}"
+        track_truncate table_name
+      end
+
+      def self.track_truncate table_name
+        Db.execute <<-EOSQL
+          update #{TRACKING_TABLE_NAME}
+          set 
+            operation = '#{operation_values[:truncate]}',
+            time = datetime('now')
+          where
+            relation_name = '#{table_name}' and
+            relation_type = '#{relation_type_values[:table]}'
+        EOSQL
       end
 
       def self.drop_table table_name
+        return if table_name.casecmp(TRACKING_TABLE_NAME) == 0
         Db.execute "drop table if exists #{table_name}"
+        track_drop table_name
+      end
+
+      def self.track_drop table_name
+        Db.execute <<-EOSQL
+          delete from #{TRACKING_TABLE_NAME} 
+          where 
+            relation_name = '#{table_name}' and 
+            relation_type = '#{relation_type_values[:table]}'
+        EOSQL
       end
 
       def self.table_exists? table_name, options = {}
@@ -75,7 +102,7 @@ module Rake
         (n_matches > 0)
       end
 
-      def self.create_table table_name, data_definition, column_definitions, track_table
+      def self.create_table table_name, data_definition, column_definitions, track_table=true
         drop_table table_name
         Db.execute <<-EOSQL
           create table #{table_name} #{column_definitions}
@@ -90,7 +117,7 @@ module Rake
       def self.operations_supported
         {
           :by_db => operations_supported_by_db,
-          :by_app => ['truncate', 'create']
+          :by_app => [:truncate, :create]
         }
       end
 
@@ -99,25 +126,27 @@ module Rake
       private
 
         def self.operations_supported_by_db
-          ['update', 'insert', 'delete']
+          [:update, :insert, :delete]
         end
 
         def self.rule_name table_name, operation
-          "#{table_name}_#{operation}"
+          "#{table_name}_#{operation.to_s}"
         end
 
         def self.create_tracking_rules table_name
           operations_supported_by_db.each do |operation|
             Db.execute <<-EOSQL
-              create trigger #{self.rule_name(table_name,operation)}
-                after #{operation} on #{table_name} begin
+              create trigger #{self.rule_name(table_name, operation)}
+                after #{operation.to_s} on #{table_name} begin
+
                   update #{TRACKING_TABLE_NAME} 
                   set 
-                    operation = '#{operation}',
-                    mtime = datetime()
+                    operation = '#{operation_values[operation]}',
+                    time = datetime()
                   where 
                     relation_name = '#{table_name}' and 
-                    relation_type = 'TABLE'
+                    relation_type = '#{relation_type_values[:table]}'
+
                 ;
                 end
             EOSQL
@@ -125,15 +154,20 @@ module Rake
         end
 
         def self.track_creation table_name, n_tuples
-          operation = 'create'
+          operation = :create
           Db.execute <<-EOSQL
             delete from #{TRACKING_TABLE_NAME} where 
               relation_name = '#{table_name}' and 
-              relation_type = 'TABLE' and
-              operation = '#{operation}'
-              ;
-            insert into #{Db::TRACKING_TABLE_NAME} values (
-              '#{table_name}', 'TABLE', '#{operation}', now()
+              relation_type = '#{relation_type_values[:table]}' and
+              operation = '#{operation_values[operation]}'
+            ;
+          EOSQL
+          Db.execute <<-EOSQL
+            insert into #{TRACKING_TABLE_NAME} values (
+              '#{table_name}', 
+              '#{relation_type_values[:table]}', 
+              '#{operation_values[operation]}', 
+              datetime('now')
             );
           EOSQL
         end
